feat: move component to the shared folder
