+++ apps/app/src/app/pages/pages/components/index.ts
@@ -1,3 +1,3 @@
 export * from './navigation-favorites/navigation-favorites.component';
 
-export * from './modal-favorites/modal-favorites.component';
+export * from '../../../shared/components/modal-favorites/modal-favorites.component';
