+++ apps/app/src/app/shared/components/index.ts
@@ -3,3 +3,5 @@ export * from './alert-error';
 export * from './auto-refresh-activator';
 
 export * from './timer-interval-selector';
+
+export * from './modal-favorites';
