+++ apps/app/src/app/shared/components/modal-favorites/index.ts
@@ -1,0 +1,1 @@
+export * from './modal-favorites.component';
