+++ apps/app/src/app/shared/components/modal-favorites/modal-favorites.component.scss
